move MapLegend to separate component
